graphite: datasource refactor
